add nats client to pubsub package, make nats settings configurable

+++ .devcontainer/scripts/nats_init.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 # script to bootstrap a nats operator environment
 
+if nsc describe operator; then
+    echo "operator exists, not overwriting config"
+    exit 0
+fi
+
 echo "Cleaning up NATS environment"
 rm -rf /nsc/*
